// apps/k-pay-app/src/types/graphql.ts
export interface Currency {
  code: string;
  name: string;
  countryCode?: string;
  precision?: number;
}
// ...
export interface Balance {
  id: string;
  amount: number;
  currency: string;
  wallet_id: string;
  updated_at: string;
}

export interface WalletTransaction {
  id: string;
  amount: string; // Decimal
  currency: Currency;
  customerPhone?: string;
  description?: string;
  exchangeRate?: string; // Decimal
  externalReference?: string;
  feeAmount?: string; // Decimal
  feeCurrency?: Currency;
  insertedAt: string; // DateTime
  provider?: string;
  providerMessage?: string;
  providerStatus?: string;
  reference?: string;
  status: string;
  transactionType: string;
  updatedAt: string; // DateTime
  wallet?: Wallet;
  counterpartyWallet?: Wallet;
}
// ...
export interface Wallet {
  id: string;
  ownerId: string;
  ownerType: string;
  status: string;
  isFrozen: boolean;
  freezeReason?: string;
  dailyLimit?: number;
  monthlyLimit?: number;
  insertedAt: string;
  updatedAt: string;
  balances: Balance[];
  transactions: WalletTransaction[];
}